Refactor explanation data and improve CSV handling; update database initialization scripts and frontend components for enhanced vocabulary management and user interaction.
