+++ .github/workflows/main_englishvocabulary.yml
@@ -28,10 +28,10 @@ jobs:
         run: |
           python -m venv venv
           source venv/bin/activate
-      
+
       - name: Install dependencies
         run: pip install -r requirements.txt
-        
+
       # Optional: Add step to run tests here (PyTest, Django test suites, etc.)
 
       - name: Zip artifact for deployment
@@ -64,7 +64,7 @@ jobs:
       - name: Unzip artifact for deployment
         run: unzip release.zip
 
-      
+
       - name: Login to Azure
         uses: azure/login@v2
         with:
@@ -78,4 +78,3 @@ jobs:
         with:
           app-name: 'EnglishVocabulary'
           slot-name: 'Production'
-
