+++ .github/workflows/pre-commit.yml
@@ -1,0 +1,27 @@
+name: pre-commit
+
+on:
+  pull_request:
+    branches:
+      - main # 保護対象ブランチ名に応じて変更
+      - develop
+jobs:
+  pre-commit:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install pre-commit
+          pre-commit install
+
+      - name: Run pre-commit
+        run: pre-commit run --all-files
