+++ .pre-commit-config.yaml
@@ -7,33 +7,46 @@ repos:
     -   id: trailing-whitespace
     -   id: end-of-file-fixer
     -   id: check-yaml
+    -   id: check-json
     -   id: check-added-large-files
+    - id: check-toml
 - repo: https://github.com/pycqa/isort
-  rev: 5.5.2
+  rev: 6.0.1
   hooks:
     - id: isort
       args: ["--profile", "black"]
 - repo: https://github.com/psf/black
-  rev: stable
+  rev: 25.1.0
   hooks:
   - id: black
     language_version: python3
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.790
+  rev: v1.15.0
   hooks:
     - id: mypy
-      args: [--ignore-missing-imports]
-- repo: https://gitlab.com/pycqa/flake8
-  rev: 3.8.1
+      args: [--ignore-missing-imports, --strict]
+- repo: https://github.com/PyCQA/flake8
+  rev: "7.2.0"
   hooks:
   - id: flake8
     # max-line-length setting is the same as black
     # commit cannot be done when cyclomatic complexity is more than 10.
-    args: [--max-line-length, "88", --ignore=E402, --max-complexity, "10", --max-expression-complexity=7, --max-cognitive-complexity=7]
+    args: [--max-line-length, "150", --ignore=E402,--ignore=E800, --max-complexity, "10", --max-expression-complexity=7, --max-cognitive-complexity=7]
     additional_dependencies: [flake8-bugbear, flake8-builtins, flake8-eradicate, pep8-naming, flake8-expression-complexity, flake8-cognitive-complexity]
-- repo: https://github.com/markdownlint/markdownlint
-  rev: master
+- repo: https://github.com/pylint-dev/pylint
+  rev: v3.3.7
   hooks:
-    - id: markdownlint
-      # ignore line length of makrdownlint
-      args: [-r, ~MD013]
+    - id: pylint
+      name: pylint
+      entry: pylint
+      language: python
+      types: [python]
+      additional_dependencies: [pandas,streamlit,langchain,dotenv,langchain_google_genai]
+      args: [--max-line-length,"150","--init-hook", "import sys; sys.path.insert(0, 'src')"]
+
+# - repo: https://github.com/markdownlint/markdownlint
+#   rev: v0.13.0
+#   hooks:
+#     - id: markdownlint
+#       # ignore line length of makrdownlint
+#       args: [-r, ~MD013]
