+++ pyproject.toml
@@ -1,0 +1,41 @@
+[build-system]
+requires = ["setuptools >= 77.0.3"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "EnglishLearnApp"
+version = "0.0.1"
+dependencies = [
+  "pandas",
+  "streamlit",
+  "langchain",
+  "dotenv",
+  "langchain_google_genai"
+]
+requires-python = ">=3.11"
+authors = [
+  {name = "Tomohito Amano", email = "amanotomohito040@gmail.com"}
+]
+maintainers = [
+  {name = "Tomohito Amano", email = "amanotomohito040@gmail.com"}
+]
+description = "Lovely Spam! Wonderful Spam!"
+readme = "README.md"
+license = "MIT"
+license-files = ["LICEN[CS]E.*"]
+keywords = ["egg", "bacon", "sausage", "tomatoes", "Lobster Thermidor"]
+classifiers = [
+  "Development Status :: 4 - Beta",
+  "Programming Language :: Python"
+]
+
+[project.urls]
+Homepage = "https://example.com"
+Documentation = "https://readthedocs.org"
+Repository = "https://github.com/me/spam.git"
+"Bug Tracker" = "https://github.com/me/spam/issues"
+Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"
+
+[tool.pylint]
+disable = ["C0114", "C0115", "C0116"]
+max-line-length = 150
