+++ requirements.txt
@@ -1,4 +1,4 @@
-langchain 
-langchain-google-genai 
+langchain
+langchain-google-genai
 pandas
 streamlit
