+++ src/backend/backend.py
@@ -1,3 +1,5 @@
+from typing import Any, List, Set, Tuple
+
 import pandas as pd
 
 from backend.core.db_core import get_db_connection, get_user_db_connection
@@ -16,6 +18,7 @@ def search_meanings(word_id: int) -> pd.DataFrame:
 
 
 def get_examples(word_id: int) -> pd.DataFrame:
+    """Get examples from word_id"""
     conn = get_db_connection()
     df = pd.read_sql_query(
         "SELECT * FROM examples WHERE word_id=?", conn, params=(word_id,)
@@ -25,6 +28,7 @@ def get_examples(word_id: int) -> pd.DataFrame:
 
 
 def is_favorited(word: str) -> bool:
+    """Check if the word is favorited"""
     conn = get_user_db_connection()
     cur = conn.cursor()
     cur.execute("SELECT 1 FROM favorites WHERE word = ?", (word,))
@@ -33,7 +37,8 @@ def is_favorited(word: str) -> bool:
     return exists
 
 
-def toggle_favorite(word: str):
+def toggle_favorite(word: str) -> None:
+    """Change the state of favorited for the word"""
     conn = get_user_db_connection()
     cur = conn.cursor()
     if is_favorited(word):
@@ -44,7 +49,8 @@ def toggle_favorite(word: str):
     conn.close()
 
 
-def get_derived_words(word_id):
+def get_derived_words(word_id: int) -> Any:
+    """Get derived words from word_id"""
     conn = get_db_connection()
     cur = conn.cursor()
     # word_id から該当 word_id を取得
@@ -54,36 +60,38 @@ def get_derived_words(word_id):
         return []
 
     # 派生語の stem_id を取得
-    cur.execute(
-        "SELECT DISTINCT stem_id FROM derived_words WHERE word_id IN ({})".format(
-            ",".join("?" * len(word_ids))
-        ),
-        word_ids,
+    placeholders = ",".join("?" for _ in word_ids)
+    query = (
+        f"SELECT DISTINCT stem_id FROM derived_words WHERE word_id IN ({placeholders})"
     )
+    cur.execute(query, word_ids)
     stem_ids = [row[0] for row in cur.fetchall()]
     if not stem_ids:
         return []
 
     # 同じ stem_id に属する他の単語（自分自身以外）を取得
-    cur.execute(
-        """
+    # プレースホルダ生成
+    placeholders = ",".join("?" for _ in stem_ids)
+
+    # SQL クエリを安全に組み立て
+    query = f"""
         SELECT DISTINCT w.word_id, w.word
         FROM words w
         JOIN derived_words d ON w.word_id = d.word_id
-        WHERE d.stem_id IN ({})
+        WHERE d.stem_id IN ({placeholders})
         AND w.word_id != ?
-        """.format(
-            ",".join("?" * len(stem_ids))
-        ),
-        stem_ids + [word_id],
-    )
+    """
+
+    # stem_ids + [word_id] をバインドパラメータとして渡す
+    cur.execute(query, stem_ids + [word_id])
     results = cur.fetchall()
     conn.close()
     return results
 
 
 # お気に入りリスト取得
-def get_favorites():
+def get_favorites() -> Any:
+    """Get all the favorited words"""
     conn = get_db_connection()
     cursor = conn.cursor()
     cursor.execute(
@@ -94,47 +102,99 @@ def get_favorites():
     return favorites
 
 
-def get_synonyms(word_id: int):
+def find_synonym_ids(start_id: int) -> set[int]:
+    """Breadth-first search over synonym graph starting from start_id."""
     conn = get_db_connection()
     cursor = conn.cursor()
+    synonym_query = """
+        SELECT word_id_2 FROM synonyms WHERE word_id_1 = ?
+        UNION
+        SELECT word_id_1 FROM synonyms WHERE word_id_2 = ?
+    """
+    visited_ids: set[int] = set()
+    to_visit = [start_id]
 
-    visited = set()
-    queue = [word_id]
-
-    while queue:
-        current = queue.pop(0)
-        if current in visited:
+    while to_visit:
+        current = to_visit.pop(0)
+        if current in visited_ids:
             continue
-        visited.add(current)
-        cursor.execute(
-            """
-            SELECT word_id_2 FROM synonyms WHERE word_id_1 = ?
-            UNION
-            SELECT word_id_1 FROM synonyms WHERE word_id_2 = ?
-        """,
-            (current, current),
-        )
+        visited_ids.add(current)
+
+        cursor.execute(synonym_query, (current, current))
         neighbors = [row[0] for row in cursor.fetchall()]
-        for neighbor in neighbors:
-            if neighbor not in visited:
-                queue.append(neighbor)
+        to_visit.extend(n for n in neighbors if n not in visited_ids)
 
-    visited.discard(word_id)  # 元の単語は除外
+    visited_ids.discard(start_id)
+    conn.close()
+    return visited_ids
 
-    # 類義語の単語を取得
-    if not visited:
-        conn.close()
+
+def fetch_word_info(word_ids: Set[int]) -> List[Tuple[int, str]]:
+    """Retrieve word_id and word text for given word_ids."""
+    conn = get_db_connection()
+    cursor = conn.cursor()
+    if not word_ids:
         return []
 
-    cursor.execute(
-        f"""
+    placeholders = ",".join("?" for _ in word_ids)
+    query = f"""
         SELECT DISTINCT w.word_id, w.word
         FROM words w
         JOIN meanings m ON w.word_id = m.word_id
-        WHERE w.word_id IN ({','.join('?' * len(visited))})
-        """,
-        list(visited),
-    )
-    results = cursor.fetchall()
+        WHERE w.word_id IN ({placeholders})
+    """
+    cursor.execute(query, list(word_ids))
+    results: List[Tuple[int, str]] = cursor.fetchall()
     conn.close()
     return results
+
+
+def get_synonyms(word_id: int) -> Any:
+    """get synonyms of given word_id"""
+    visited_ids = find_synonym_ids(word_id)
+    results = fetch_word_info(visited_ids)
+    return results
+
+    # conn = get_db_connection()
+    # cursor = conn.cursor()
+
+    # visited: Set[int] = set()
+    # queue = [word_id]
+
+    # while queue:
+    #     current = queue.pop(0)
+    #     if current in visited:
+    #         continue
+    #     visited.add(current)
+    #     cursor.execute(
+    #         """
+    #         SELECT word_id_2 FROM synonyms WHERE word_id_1 = ?
+    #         UNION
+    #         SELECT word_id_1 FROM synonyms WHERE word_id_2 = ?
+    #         """,
+    #         (current, current),
+    #     )
+    #     neighbors = [row[0] for row in cursor.fetchall()]
+    #     for neighbor in neighbors:
+    #         if neighbor not in visited:
+    #             queue.append(neighbor)
+
+    # visited.discard(word_id)  # 元の単語は除外
+
+    # # 類義語の単語を取得
+    # if not visited:
+    #     conn.close()
+    #     return []
+
+    # cursor.execute(
+    #     f"""
+    #     SELECT DISTINCT w.word_id, w.word
+    #     FROM words w
+    #     JOIN meanings m ON w.word_id = m.word_id
+    #     WHERE w.word_id IN ({','.join('?' * len(visited))})
+    #     """,
+    #     list(visited),
+    # )
+    # results = cursor.fetchall()
+    # conn.close()
+    # return results
