+++ src/data/explanation_data/lv12_explanation.csv
@@ -8726,23 +8726,23 @@ word_id,word,explanation
 
 ### 🔹 ニュアンス
 
-- 主に**航空機**を格納・整備するための**大きな建造物**を指す  
-- 通常は空港や軍の基地にあり、**屋内で保護・修理を行う目的**で使用される  
+- 主に**航空機**を格納・整備するための**大きな建造物**を指す
+- 通常は空港や軍の基地にあり、**屋内で保護・修理を行う目的**で使用される
 - 鉄骨や金属製の頑丈な構造が多く、**shed や garage** とは異なり、**産業的なスケール感**がある
 
 #### 例文
 
-- The aircraft was moved into the hangar for maintenance.  
-- A massive hangar housed several military jets.  
-- The storm damaged the roof of the old hangar.  
+- The aircraft was moved into the hangar for maintenance.
+- A massive hangar housed several military jets.
+- The storm damaged the roof of the old hangar.
 
 ---
 
 ## 🌱 語源
 
-- フランス語 **hangar**（納屋、小屋）  
-- 中期フランス語 **hanghart**（囲い、囲まれた場所）が語源  
-- 元々は「保護されたスペース」という意味から派生  
+- フランス語 **hangar**（納屋、小屋）
+- 中期フランス語 **hanghart**（囲い、囲まれた場所）が語源
+- 元々は「保護されたスペース」という意味から派生
 
 ---
 
@@ -10847,27 +10847,27 @@ word_id,word,explanation
 ## ✅ 単語の意味やニュアンス
 
 - **品詞**: 他動詞 (transitive verb)
-- **意味**:  
-  - （人を）隔離する、人里離れた場所に引きこもらせる  
+- **意味**:
+  - （人を）隔離する、人里離れた場所に引きこもらせる
   - 社会的または物理的に**意図的に切り離す**
 
 ### 🔹 ニュアンス
 
-- 「自分（または誰か）を他人の目から**遠ざける**」「隠す」ニュアンス  
-- 主に**静かさ・プライバシー・安全**などを求めて、自発的または他人によって隔てられる行為  
-- **精神的な孤立**や、修道士のような**隠遁生活**にも用いられる  
+- 「自分（または誰か）を他人の目から**遠ざける**」「隠す」ニュアンス
+- 主に**静かさ・プライバシー・安全**などを求めて、自発的または他人によって隔てられる行為
+- **精神的な孤立**や、修道士のような**隠遁生活**にも用いられる
 
 #### 例文
 
-- She secluded herself in the mountains to finish her novel.  
-- The monks live in a secluded monastery.  
-- The cabin was secluded deep in the woods.  
+- She secluded herself in the mountains to finish her novel.
+- The monks live in a secluded monastery.
+- The cabin was secluded deep in the woods.
 
 ---
 
 ## 🌱 語源
 
-- ラテン語 **secludere**（引き離す）  
+- ラテン語 **secludere**（引き離す）
   - **se-**（離れて）+ **claudere**（閉じる）→ 「閉じて隔てる」
 
 ---
@@ -36912,15 +36912,15 @@ jubilee は祝祭を意味するため、直接的な対義語は存在しにく
 
 - **品詞**: 他動詞 (transitive verb)
 - **意味**:
-  1. （腰に）帯・ベルトを巻く、装着する  
-  2. （周囲を）取り囲む、取り巻く  
+  1. （腰に）帯・ベルトを巻く、装着する
+  2. （周囲を）取り囲む、取り巻く
   3. **《gird oneself》心構えをする、気を引き締める**
 
 ### 🔹 ニュアンス
 
-- 本来は「帯で囲む、巻く」などの**物理的な行為**を意味するが、現代英語では特に比喩的に「気持ちを整える」「備える」という意味で多く使われる  
-- 古風あるいは文語的表現として扱われることが多い  
-- 特に **困難や戦いに向けての精神的準備** を表す表現として文学・スピーチなどで使われる  
+- 本来は「帯で囲む、巻く」などの**物理的な行為**を意味するが、現代英語では特に比喩的に「気持ちを整える」「備える」という意味で多く使われる
+- 古風あるいは文語的表現として扱われることが多い
+- 特に **困難や戦いに向けての精神的準備** を表す表現として文学・スピーチなどで使われる
 
 #### 例文
 
@@ -36934,9 +36934,9 @@ jubilee は祝祭を意味するため、直接的な対義語は存在しにく
 
 ## 🌱 語源
 
-- 古英語 **'gyrdan'** = 包む、帯を巻く  
-- ゲルマン系語源であり、ドイツ語 **gürten**（帯を締める）と同根  
-- 転じて、「心を締める」＝「備える」という比喩的意味に派生  
+- 古英語 **'gyrdan'** = 包む、帯を巻く
+- ゲルマン系語源であり、ドイツ語 **gürten**（帯を締める）と同根
+- 転じて、「心を締める」＝「備える」という比喩的意味に派生
 
 ---
