+++ src/data/generate_synonym/lv7_synonyms.csv
@@ -48510,7 +48510,7 @@ number,word,synonym
 6342,competence,hitch
 6342,competence,catch
 6342,competence,rub
-6342,competence,Impediment 
+6342,competence,Impediment
 6343,accompaniment,support
 6343,accompaniment,backing
 6343,accompaniment,complement
