+++ src/data/generate_synonym/lv8_synonyms.csv
@@ -36116,7 +36116,7 @@ number,word,synonym
 7170,shooting,flinging
 7170,shooting,pitching
 7170,shooting,slinging
-7170,shooting,catapulting 
+7170,shooting,catapulting
 7171,windmill,wind turbine
 7171,windmill,wind power plant
 7171,windmill,wind generator
