+++ src/data/generate_synonym/lv9_synonyms.csv
@@ -32613,7 +32613,7 @@ number,word,synonym
 8329,tabloid,supermarket tabloid
 8329,tabloid,scurrilous publication
 8329,tabloid,downmarket newspaper
-8329,tabloid, Fleet Street 
+8329,tabloid, Fleet Street
 8329,tabloid,broadsheet (ironically)
 8329,tabloid,penny dreadful (historical)
 8329,tabloid,"journal of record (antonym, ironically)"
